+++ integration/jest.config.js
@@ -3,8 +3,9 @@ module.exports = {
   globalSetup: './setup.js',
   globalTeardown: './teardown.js',
   setupTestFrameworkScriptFile: undefined,
+  testEnvironment: './testEnvironment.js',
   preset: 'jest-puppeteer-docker',
-  testRegex: './*\\.test\\.ts$',
+  testRegex: './*\\.test\\.(?:t|j)s$',
   moduleFileExtensions: ['ts', 'js'],
   transform: {
     '^.+\\.(ts|tsx?)$': 'ts-jest',