+++ src/components/Headline/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import styled, { themed } from '../../utils/styled';
-import { TextStyles, TextStylings } from '../../textStyles';
+import styled, { themed, css } from '../../utils/styled';
 import { StandardProps } from '../../common';
 import { distance } from '../../distance';
+import { getFontStyle } from '../../textStyles';
 
 export type HeadlineSize = 'small' | 'medium';
 
@@ -13,49 +13,20 @@ export interface HeadlineProps extends StandardProps {
    */
   level?: 1 | 2 | 3 | 4 | 5;
   /**
-   * Currently supports 2 different sizes, SMALL and MEDIUM,
-   * which could be used for ex. responsive layouts
-   */
-  size?: HeadlineSize;
-  /**
    * When specified headline will have muted text color
    */
   subheader?: boolean;
-  /**
-   * When specified will select provided text style
-   */
-  textStyle?: TextStyles;
   /**
    * The text of the headline.
    */
   children?: React.ReactNode;
 }
 
-const sizeMapping = {
-  medium: {
-    h1: TextStyles.giga,
-    h2: TextStyles.mega,
-    h3: TextStyles.alpha,
-    h4: TextStyles.beta,
-    h5: TextStyles.gamma,
-    h6: TextStyles.delta,
-  },
-  small: {
-    h1: TextStyles.alpha,
-    h2: TextStyles.beta,
-    h3: TextStyles.gamma,
-    h4: TextStyles.delta,
-    h5: TextStyles.epsilon,
-    h6: TextStyles.zeta,
-  },
-};
-
 export interface StyledHeadlineProps {
   size: HeadlineSize;
   level: number;
   theme?: any;
   subheader?: boolean;
-  textStyle?: TextStyles;
 }
 
 interface HeadlineCache {
@@ -65,14 +36,24 @@ interface HeadlineCache {
 const Headlines: HeadlineCache = {};
 
 function getComponentName(level: number) {
-  return `h${level >= 1 && level <= 6 ? level : 3}`;
+  return `h${level >= 1 && level <= 5 ? level : 3}`;
 }
 
-function getTextStyling(props: StyledHeadlineProps, key: string): string {
-  const { size, level } = props;
-  const component = getComponentName(level);
-  const textStyling = props.textStyle ? TextStylings[props.textStyle] : TextStylings[sizeMapping[size][component]];
-  return textStyling[key];
+function getHeadlineStyle(level: StyledHeadlineProps['level']) {
+  switch (level) {
+    case 1:
+      return getFontStyle({ size: 'xxxLarge', weight: 'light' });
+    case 2:
+      return getFontStyle({ size: 'xxLarge', weight: 'light' });
+    case 3:
+      return getFontStyle({ size: 'xLarge', weight: 'medium' });
+    case 4:
+      return getFontStyle({ size: 'large', weight: 'regular' });
+    case 5:
+      return getFontStyle({ size: 'medium', weight: 'bold' });
+    default:
+      return '';
+  }
 }
 
 function getStyledHeadline(level: number) {
@@ -80,11 +61,18 @@ function getStyledHeadline(level: number) {
   const Headline = Headlines[component];
 
   if (!Headline) {
-    const NewHeadline = styled[component]`
-      margin: 0;
-      font-weight: inherit;
-      font-size: inherit;
-    `;
+    const NewHeadline = styled(component as 'h1')<StyledHeadlineProps>(
+      themed(
+        props => css`
+          ${getHeadlineStyle(props.level)}
+
+          margin: 0;
+          padding: ${props.theme.headingsPadding || `0 ${distance.small} 0 0`};
+          font-family: ${props.theme.fontFamily || 'inherit'};
+          color: ${props.subheader ? props.theme.text5 : 'inherit'};
+        `,
+      ),
+    );
     Headlines[component] = NewHeadline;
     return NewHeadline;
   }
@@ -92,27 +80,16 @@ function getStyledHeadline(level: number) {
   return Headline;
 }
 
-const HeadlineContainer = styled('div')<StyledHeadlineProps>`
-  padding: ${themed(props => props.theme.headingsPadding || `0 ${distance.small} 0 0`)};
-  font-size: ${props => getTextStyling(props, 'fontSize')};
-  font-weight: ${props => getTextStyling(props, 'fontWeight')};
-  text-transform: ${props => getTextStyling(props, 'textTransform')};
-  font-family: ${themed(props => props.theme.fontFamily || 'inherit')};
-  line-height: ${props => getTextStyling(props, 'lineHeight')};
-  letter-spacing: ${props => getTextStyling(props, 'letterSpacing')};
-  color: ${themed(props => (props.subheader ? props.theme.text5 : 'inherit'))};
-`;
-
-const defaultSize: HeadlineSize = 'medium';
 /**
  * Headline component with styles for all headline levels.
  */
-export const Headline: React.SFC<HeadlineProps> = ({ level = 3, size = defaultSize, children, ...rest }) => {
+export const Headline: React.SFC<HeadlineProps> = ({ level = 3, children, ...rest }) => {
   const StyledHeadline = getStyledHeadline(level);
   return (
-    <HeadlineContainer level={level} size={size} {...rest}>
-      <StyledHeadline>{children}</StyledHeadline>
-    </HeadlineContainer>
+    <StyledHeadline level={level} {...rest}>
+      {children}
+    </StyledHeadline>
   );
 };
+
 Headline.displayName = 'Headline';