+++ src/components/MetroInfoTile/index.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import styled, { themed } from '../../utils/styled';
 import { Anchor, AnchorProps } from '../Anchor';
 import { StyledTileIcon } from '../../quarks';
+import { getFontStyle } from '../../textStyles';
 
 const Tile = styled(Anchor)`
   cursor: pointer;
@@ -21,13 +22,13 @@ const Tile = styled(Anchor)`
 `;
 
 const TileTitle = styled.span`
+  ${getFontStyle({ size: 'small', weight: 'medium' })}
+
   position: absolute;
   left: 0;
   bottom: 0;
   right: 0;
   padding: 0 10px 10px;
-  font-size: 0.875rem;
-  font-weight: 500;
   overflow: hidden;
   white-space: nowrap;
   text-overflow: ellipsis;