+++ src/components/TabPanel/TabPanel.part.tsx
@@ -2,7 +2,7 @@ import styled, { css, themed } from '../../utils/styled';
 import { StandardProps } from '../../common';
 import { dark, cyan } from '../../colors';
 import { distance } from '../../distance';
-import { TextStylings } from '../../textStyles';
+import { getFontStyle } from '../../textStyles';
 
 export const TabContainer = styled.div``;
 
@@ -15,8 +15,9 @@ export const TabHeaders = styled.ul`
 export const TabContent = styled.div``;
 
 const ActiveTab = css`
+  ${getFontStyle({ weight: 'medium' })}
+
   color: ${themed(props => props.theme.text6)};
-  font-weight: 500;
   border-color: ${dark};
 `;
 
@@ -32,6 +33,7 @@ export interface TabHeaderProps extends StandardProps {
 }
 
 export const TabHeader = styled('li')<TabHeaderProps>`
+  ${getFontStyle({ size: 'large', weight: 'regular' })}
   position: relative;
   z-index: 1;
   margin: 0;
@@ -40,9 +42,6 @@ export const TabHeader = styled('li')<TabHeaderProps>`
   overflow: visible;
   text-overflow: ellipsis;
   white-space: nowrap;
-  line-height: ${TextStylings.delta.lineHeight};
-  font-size: ${TextStylings.delta.fontSize};
-  font-weight: 400;
   box-sizing: border-box;
   border-bottom: 2px solid transparent;
   margin-right: ${distance.xxlarge};