+++ tools/icongen.config.js
@@ -87,6 +87,10 @@ module.exports = {
     'ic_unfold_less_24px.svg',
     'ic_fullscreen_24px.svg',
     'ic_fullscreen_exit_24px.svg',
+    'ic_power_settings_new_24px.svg',
+    'ic_language_24px.svg',
+    'ic_sort_24px.svg',
+    'ic_pageview_24px.svg',
   ],
   out: 'src/components/Icon',
 };