+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Precise UI Changelog
 
+## 0.5.1
+
+- `Flyout` fixed compatibility with IE11
+- `Headline` styles update
+- Updated label in `Modal`
+
 ## 0.5.0
 
 - Initial release